refactor: refactored the whole python data pipeline

+++ data_pipeline/script/sql/creation/create_news_table.sql
@@ -5,6 +5,6 @@ CREATE TABLE IF NOT EXISTS news (
     link TEXT UNIQUE NOT NULL,
     lang TEXT,
     full_text TEXT,
-    content_hash CHAR(64) NOT NULL UNIQUE
-    source_id integer REFERENCES source(id)
+    content_hash CHAR(64) NOT NULL UNIQUE,
+    source_id BIGINT REFERENCES source(id)
 );
